Fixes #9136
Delayed the deletion of null keys within coalesce until after
subdependencies have been processed. The behaviour
previously was that a subcharts null keys would be processed
in a parent chart prior to the dependency being processed.
This then caused the child charts keys to be merged back
into the parent values.

Signed-off-by: James Payne <jamoflaw@gmail.com>