Further delaying the removal of null keys from the coalesced values.
Split the Coalesce* functions into KeepNil and TrimNil
variants to be clear to their intention. The issue stems
from the Coalesce functions not be idempotent when
it comes to null removal and they are called
continuously in the render pipeline.

The TrimNil variants are identical to the original
Coalesce* functions and the original public functions
point back to the renamed variants.

Signed-off-by: James Payne <jamoflaw@gmail.com>